refactor(dataset): add flag to relax criteria to call Biogrid API to accommodate (#14)
