+++ src/data/generate_positive_dataset.py
@@ -15,6 +15,7 @@ import aiohttp
 
 sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), "../../")))
 from core import config as cfg
+from src.data.bio_apis import get_interactors
 from src.data.data_processing import parse_input_genes, remove_ground_truth_data
 
 
@@ -32,81 +33,6 @@ def write_ppi_file(ppi_list, outfile):
         writer.writerows(rows)
 
 
-async def get_interactors(
-    session: aiohttp.ClientSession,
-    gene_list: list, 
-    cross_study_level: int = 2
-) -> list:
-    """
-    Retrive a dataset of all interactors for all genes
-    of interest.
-
-    Parameters
-    ----------
-    session: aiohttp.ClientSession
-        The current asynchronous session to make the async 
-        calls to the BioGrid API.
-    gene_list: list(str)
-        A list of the official symbols of a gene
-    cross_study_level: int
-        The minimum number of different studies that have 
-        confirmed a PPI. Default is 2, and it is recommended
-        that this value not be set to below 2
-    """
-    request_url = cfg.BIOGRID_BASE_URL + "/interactions"
-    evidence_list = [
-        "affinity capture-luminescence",
-        "affinity capture-ms",
-        "affinity capture-rna",
-        "affinity capture-western",
-        "two-hybrid"
-    ]
-    params = {
-        "accesskey": cfg.BIOGRID_API_KEY,
-        "format": "json",
-        "geneList": "|".join(gene_list),
-        "interSpeciesExcluded": "true",
-        "selfInteractionsExcluded": "true",
-        "evidenceList": "|".join(evidence_list),
-        "includeEvidence": "true",
-        "includeInteractors": "true",
-        "searchNames": "true",
-        "throughputTag": "low",
-        "taxId": 9606, # Homo Sapiens taxonomy ID
-        "includeHeader": "true"
-    }
-    resp = await session.request('GET', url=request_url, params=params)
-    if resp.status != 200:
-        logging.warning("Failed to get request for one of the genes in the gene list")
-        return
-    
-    interactions = await resp.json()
-    if len(interactions) == 0:
-        logging.warning("Failed to retrieve any interaction experiments for the genes in the gene list")
-        return
-    
-    # Create a hashmap of PPI pairs and the number of unique experiments. Unique experiments
-    # are experiments that are from different studies (different Pubmed IDs).
-    ppis = {}
-    for interaction in interactions.values():
-        sym_a = interaction['OFFICIAL_SYMBOL_A']
-        sym_b = interaction['OFFICIAL_SYMBOL_B']
-        # Ensure pair is not already in the dataset
-        experimental_id = f'{interaction["PUBMED_ID"]}_{interaction["EXPERIMENTAL_SYSTEM"]}'
-        if f'{sym_a}_{sym_b}' in ppis:
-            ppis[f'{sym_a}_{sym_b}'].add(experimental_id)
-        elif f'{sym_b}_{sym_a}' in ppis:
-            ppis[f'{sym_b}_{sym_a}'].add(experimental_id)
-        # Write cancer driver gene first
-        elif sym_a in gene_list:
-            ppis[f'{sym_a}_{sym_b}'] = {experimental_id}
-        else: # Gene b in gene_list
-            ppis[f'{sym_b}_{sym_a}'] = {experimental_id}
-    high_conf_ppis = [ppi for ppi, e_id in ppis.items() if len(e_id) >= cross_study_level]
-    logging.debug(f'Found a total of {len(interactions)} and generated {len(high_conf_ppis)} high confidence PPIs...')
-    return high_conf_ppis
-
-
 def chunk_input_genes(input_genes: list, chunk_size: int = 60) -> list:
     """Chunk input genes since the Biogrid API is limited to returning
     10,000 interactions."""
@@ -120,7 +46,7 @@ def parse_command_line(): # pragma: no cover
     parser.add_argument('infile', type=str,
                         help='Path to Input CSV containing genes of interest')
     parser.add_argument('outfile', type=str,
-                        help='Output file path to write list of genes to')
+                        help='Output file path to write positive dataset to')
     parser.add_argument('-v', '--verbose',
                         action='store_true',
                         help='Change logging level from default level to noisiest level')
